src/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
requests==2.32.5
pandas==2.3.3
google-cloud-storage==3.7.0
gcsfs==2025.12.0